+++ cryptomonnaie/src/main/java/com/cda/jee/dao/HoldingsDAOImp.java
@@ -24,10 +24,10 @@ public class HoldingsDAOImp implements IDAO<Holdings>{
 			ResultSet r = st.executeQuery("select * from holding natural join currency");
 			while (r.next()) {
 				res.add(new Holdings(r.getInt("id_holding"),r.getString("name_currency"),r.getInt("quantity"),
-						r.getFloat("unit_purchase_price"),r.getDate("purchase_date"),r.getFloat("current_price"),r.getFloat("current_price")-r.getFloat("unit_purchase_price")));
+						r.getFloat("unit_purchase_price"),r.getDate("purchase_date"),r.getFloat("current_price"),r.getFloat("unit_purchase_price")-r.getFloat("current_price")));
 			}
 		} catch (SQLException e) {
-			//log + erreur à faire
+			logger.error("erreur : " + e);
 		}
 		return res;
 	}
@@ -53,13 +53,13 @@ public class HoldingsDAOImp implements IDAO<Holdings>{
 					}
 					return t;
 				} catch (SQLException e) {
-					//log + erreur à faire + return t à virer
-					return t;
+					logger.error("erreur : " + e);
+					return null;
 				}
 			}
 			catch (SQLException e) {
-				//log + erreur à faire + return t à virer
-				return t;			
+				logger.error("erreur : " + e);	
+				return null;	
 			}
 		}
 		else {
@@ -81,8 +81,7 @@ public class HoldingsDAOImp implements IDAO<Holdings>{
 						r.getFloat("unit_purchase_price"),r.getDate("purchase_date"),r.getFloat("current_price"),r.getFloat("current_price")-r.getFloat("unit_purchase_price"));
 			}
 		} catch (SQLException e) {
-			//log + erreur à faire + return res à virer
-			return res;
+			logger.error("erreur : " + e);
 		}
 		return res;
 	}
@@ -90,36 +89,48 @@ public class HoldingsDAOImp implements IDAO<Holdings>{
 	@Override
 	public Holdings updateById(Holdings t) {
 		Connection c = MyConnection.getConnection();
-		try(PreparedStatement statement = c.prepareStatement("select id_currency from currency where currency_name = ?;")){
-			statement.setString(1, t.getNameCurrency());
-			ResultSet r = statement.executeQuery();
-			try (PreparedStatement ps = c.prepareStatement("update holding set purchase_date=?, unit_purchase_price=?,quantity=?,id_currency=? where id_holding=?")) {
-				ps.setDate(1, t.getPurchaseDate());
-				ps.setFloat(2, t.getPurchasePrice());
-				ps.setInt(3, t.getQuantity());
-				ps.setInt(4, r.getInt("id_currency"));
-				ps.executeUpdate();
-				return t;
-			} catch (SQLException e) {
-				//log + erreur à faire + return t à virer
-				return t;
+		Holdings holding = getByName(t.getNameCurrency());
+		if(holding!=null) {
+			try(PreparedStatement statement = c.prepareStatement("select id_currency from currency where currency_name = ?;")){
+				statement.setString(1, t.getNameCurrency());
+				ResultSet r = statement.executeQuery();
+				try (PreparedStatement ps = c.prepareStatement("update holding set purchase_date=?, unit_purchase_price=?,quantity=?,id_currency=? where id_holding=?")) {
+					ps.setDate(1, t.getPurchaseDate());
+					ps.setFloat(2, t.getPurchasePrice());
+					ps.setInt(3, t.getQuantity());
+					ps.setInt(4, r.getInt("id_currency"));
+					ps.setInt(5, t.getIdHoldings());
+					ps.executeUpdate();
+					return t;
+				} catch (SQLException e) {
+					logger.error("erreur : " + e);
+				}
+			}
+			catch (SQLException e) {
+				logger.error("erreur : " + e);	
 			}
 		}
-		catch (SQLException e) {
-			//log + erreur à faire + return t à virer
-			return t;			
+		else {
+			logger.error("erreur : l'avoir n'existe pas en BDD");
 		}
+		return t;
 	}
 
 	@Override
 	public void deleteById(int id) {
 		Connection c = MyConnection.getConnection();
-		try (PreparedStatement ps = c.prepareStatement("delete from holding where id_holding=?")) {
-			ps.setInt(1, id);
-			ps.executeUpdate();
-		} 
-		catch (SQLException e) {
-			//log + erreur à faire
+		Holdings holding = getById(id);
+		if(holding!=null) {
+			try (PreparedStatement ps = c.prepareStatement("delete from holding where id_holding=?")) {
+				ps.setInt(1, id);
+				ps.executeUpdate();
+			} 
+			catch (SQLException e) {
+				logger.error("erreur : " + e);
+			}
+		}
+		else {
+			logger.error("erreur : l'avoir n'existe pas en BDD");			
 		}
 	}
 
@@ -135,8 +146,7 @@ public class HoldingsDAOImp implements IDAO<Holdings>{
 						r.getFloat("unit_purchase_price"),r.getDate("purchase_date"),r.getFloat("current_price"),r.getFloat("current_price")-r.getFloat("unit_purchase_price"));
 			}
 		} catch (SQLException e) {
-			//log + erreur à faire + return res à virer
-			return res;
+			logger.error("erreur : " + e);
 		}
 		return res;
 	}
